Bug Fixed in Edit Staff Components.

+++ src/component/ManageStaff.jsx
@@ -17,11 +17,13 @@ export default function ManageStaff() {
       const res = await axios.get(
         `${import.meta.env.VITE_APP_BASE_URL}/user/getAllUser`
       );
+      console.log(res, "resposen.data");
       const receptionists = res.data.user.filter(
         (user) =>
           user.role === "receptionist" &&
           user.branchId === localStorage.getItem("selectedBranch")
       );
+      console.log(receptionists, "receptionists");
       setStaffList(receptionists);
     } catch (error) {
       console.error("Error fetching staff:", error);
@@ -42,6 +44,9 @@ export default function ManageStaff() {
   }, []);
 
   const handleDelete = async (id) => {
+    const confirmDelete = window.confirm("Are you sure you want to delete this staff?");
+    if (!confirmDelete) return;
+  
     try {
       await axios.delete(
         `${import.meta.env.VITE_APP_BASE_URL}/user/deleteUserById/${id}`
